requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
regex==2025.7.34
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
